+++ beaker/services/workspace.py
@@ -1,8 +1,10 @@
 from collections import defaultdict
-from typing import Dict, List, Optional, Union
+from datetime import datetime
+from typing import Dict, Generator, List, Optional, Union
 
 from ..data_model import *
 from ..exceptions import *
+from ..util import format_cursor
 from .service_client import ServiceClient
 
 
@@ -51,7 +53,7 @@ class WorkspaceClient(ServiceClient):
             raise
 
     def create(
-        self, workspace: str, description: Optional[str] = None, public: bool = False
+        self, workspace: str, *, description: Optional[str] = None, public: bool = False
     ) -> Workspace:
         """
         Create a workspace.
@@ -209,36 +211,25 @@ class WorkspaceClient(ServiceClient):
             },
         )
 
-    def list(
+    def iter(
         self,
         org: Optional[Union[str, Organization]] = None,
+        *,
         author: Optional[Union[str, Account]] = None,
         match: Optional[str] = None,
         archived: Optional[bool] = None,
         limit: Optional[int] = None,
-    ) -> List[Workspace]:
-        """
-        List workspaces belonging to an organization.
-
-        :param org: The organization name or object. If not specified,
-            :data:`Beaker.config.default_org <beaker.Config.default_org>` is used.
-        :param author: Only list workspaces authored by this account.
-        :param match: Only include workspaces matching the text.
-        :param archived: Only include/exclude archived workspaces.
-        :param limit: Limit the number of workspaces returned.
-
-        :raises OrganizationNotFound: If the organization doesn't exist.
-        :raises OrganizationNotSet: If neither ``org`` nor
-            :data:`Beaker.config.default_org <beaker.Config.default_org>` are set.
-        :raises AccountNotFound: If the author account doesn't exist.
-        :raises BeakerError: Any other :class:`~beaker.exceptions.BeakerError` type that can occur.
-        :raises RequestException: Any other exception that can occur when contacting the
-            Beaker server.
-        """
+        sort_by: WorkspaceSort = WorkspaceSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> Generator[Workspace, None, None]:
         org = self.resolve_org(org)
-        workspaces: List[Workspace] = []
-        cursor: Optional[str] = None
-        query: Dict[str, str] = {"org": org.id}
+        query: Dict[str, str] = {
+            "org": org.id,
+            "field": str(sort_by),
+            "order": "descending" if descending else "ascending",
+            "cursor": format_cursor(cursor),
+        }
         if author is not None:
             query["author"] = (
                 author.name if isinstance(author, Account) else self.beaker.account.get(author).name
@@ -250,8 +241,8 @@ class WorkspaceClient(ServiceClient):
         if limit:
             query["limit"] = str(limit)
 
+        count = 0
         while True:
-            query["cursor"] = cursor or ""
             page = WorkspacePage.from_json(
                 self.request(
                     "workspaces",
@@ -259,29 +250,82 @@ class WorkspaceClient(ServiceClient):
                     query=query,
                 ).json()
             )
-            workspaces.extend(page.data)
-            cursor = page.next_cursor
-            if not cursor:
-                break
-            if limit and len(workspaces) >= limit:
-                workspaces = workspaces[:limit]
+            for workspace in page.data:
+                count += 1
+                yield workspace
+                if limit is not None and count >= limit:
+                    return
+
+            query["cursor"] = page.next_cursor or page.next  # type: ignore
+            if not query["cursor"]:
                 break
 
-        return workspaces
+    def list(
+        self,
+        org: Optional[Union[str, Organization]] = None,
+        *,
+        author: Optional[Union[str, Account]] = None,
+        match: Optional[str] = None,
+        archived: Optional[bool] = None,
+        limit: Optional[int] = None,
+        sort_by: WorkspaceSort = WorkspaceSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> List[Workspace]:
+        """
+        List workspaces belonging to an organization.
 
-    def images(
+        :param org: The organization name or object. If not specified,
+            :data:`Beaker.config.default_org <beaker.Config.default_org>` is used.
+        :param author: Only list workspaces authored by this account.
+        :param match: Only include workspaces matching the text.
+        :param archived: Only include/exclude archived workspaces.
+        :param limit: Limit the number of workspaces returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
+
+        :raises OrganizationNotFound: If the organization doesn't exist.
+        :raises OrganizationNotSet: If neither ``org`` nor
+            :data:`Beaker.config.default_org <beaker.Config.default_org>` are set.
+        :raises AccountNotFound: If the author account doesn't exist.
+        :raises BeakerError: Any other :class:`~beaker.exceptions.BeakerError` type that can occur.
+        :raises RequestException: Any other exception that can occur when contacting the
+            Beaker server.
+        """
+        return list(
+            self.iter(
+                org=org,
+                author=author,
+                match=match,
+                archived=archived,
+                limit=limit,
+                sort_by=sort_by,
+                descending=descending,
+                cursor=cursor,
+            )
+        )
+
+    def iter_images(
         self,
         workspace: Optional[Union[str, Workspace]] = None,
+        *,
         match: Optional[str] = None,
         limit: Optional[int] = None,
-    ) -> List[Image]:
+        sort_by: ImageSort = ImageSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> Generator[Image, None, None]:
         """
-        List the images in a workspace.
+        Iterate over the images in a workspace.
 
         :param workspace: The Beaker workspace name or object. If not specified,
             :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
         :param match: Only include images matching the text.
         :param limit: Limit the number of images returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
 
         :raises WorkspaceNotFound: If the workspace doesn't exist.
         :raises WorkspaceNotSet: If neither ``workspace`` nor
@@ -291,14 +335,18 @@ class WorkspaceClient(ServiceClient):
             Beaker server.
         """
         workspace_name = self.resolve_workspace(workspace, read_only_ok=True).full_name
-        images: List[Image] = []
-        cursor: Optional[str] = None
-        query: Dict[str, str] = {}
+        query: Dict[str, str] = {
+            "field": str(sort_by),
+            "order": "descending" if descending else "ascending",
+            "cursor": format_cursor(cursor),
+        }
         if match is not None:
             query["q"] = match
+        if limit:
+            query["limit"] = str(limit)
 
+        count = 0
         while True:
-            query["cursor"] = cursor or ""
             page = ImagesPage.from_json(
                 self.request(
                     f"workspaces/{self.url_quote(workspace_name)}/images",
@@ -309,29 +357,76 @@ class WorkspaceClient(ServiceClient):
                     },
                 ).json()
             )
-            images.extend(page.data)
-            cursor = page.next_cursor
-            if not cursor:
-                break
-            if limit is not None and len(images) >= limit:
-                images = images[:limit]
+
+            for image in page.data:
+                count += 1
+                yield image
+                if limit is not None and count >= limit:
+                    return
+
+            query["cursor"] = page.next_cursor or page.next  # type: ignore
+            if not query["cursor"]:
                 break
 
-        return images
-
-    def experiments(
+    def images(
         self,
         workspace: Optional[Union[str, Workspace]] = None,
+        *,
         match: Optional[str] = None,
         limit: Optional[int] = None,
-    ) -> List[Experiment]:
+        sort_by: ImageSort = ImageSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> List[Image]:
         """
-        List the experiments in a workspace.
+        List the images in a workspace.
+
+        :param workspace: The Beaker workspace name or object. If not specified,
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
+        :param match: Only include images matching the text.
+        :param limit: Limit the number of images returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
+
+        :raises WorkspaceNotFound: If the workspace doesn't exist.
+        :raises WorkspaceNotSet: If neither ``workspace`` nor
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` are set.
+        :raises BeakerError: Any other :class:`~beaker.exceptions.BeakerError` type that can occur.
+        :raises RequestException: Any other exception that can occur when contacting the
+            Beaker server.
+        """
+        return list(
+            self.iter_images(
+                workspace=workspace,
+                match=match,
+                limit=limit,
+                sort_by=sort_by,
+                descending=descending,
+                cursor=cursor,
+            )
+        )
+
+    def iter_experiments(
+        self,
+        workspace: Optional[Union[str, Workspace]] = None,
+        *,
+        match: Optional[str] = None,
+        limit: Optional[int] = None,
+        sort_by: ExperimentSort = ExperimentSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> Generator[Experiment, None, None]:
+        """
+        Iterate over the experiments in a workspace.
 
         :param workspace: The Beaker workspace name or object. If not specified,
             :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
         :param match: Only include experiments matching the text.
         :param limit: Limit the number of experiments returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
 
         :raises WorkspaceNotFound: If the workspace doesn't exist.
         :raises WorkspaceNotSet: If neither ``workspace`` nor
@@ -341,14 +436,18 @@ class WorkspaceClient(ServiceClient):
             Beaker server.
         """
         workspace_name = self.resolve_workspace(workspace, read_only_ok=True).full_name
-        experiments: List[Experiment] = []
-        cursor: Optional[str] = None
-        query: Dict[str, str] = {}
+        query: Dict[str, str] = {
+            "field": str(sort_by),
+            "order": "descending" if descending else "ascending",
+            "cursor": format_cursor(cursor),
+        }
         if match is not None:
             query["q"] = match
+        if limit:
+            query["limit"] = str(limit)
 
+        count = 0
         while True:
-            query["cursor"] = cursor or ""
             page = ExperimentsPage.from_json(
                 self.request(
                     f"workspaces/{self.url_quote(workspace_name)}/experiments",
@@ -359,26 +458,70 @@ class WorkspaceClient(ServiceClient):
                     },
                 ).json()
             )
-            experiments.extend(page.data)
-            cursor = page.next_cursor
-            if not cursor:
-                break
-            if limit is not None and len(experiments) >= limit:
-                experiments = experiments[:limit]
+
+            for experiment in page.data:
+                count += 1
+                yield experiment
+                if limit is not None and count >= limit:
+                    return
+
+            query["cursor"] = page.next_cursor or page.next  # type: ignore
+            if not query["cursor"]:
                 break
 
-        return experiments
-
-    def datasets(
+    def experiments(
         self,
         workspace: Optional[Union[str, Workspace]] = None,
+        *,
+        match: Optional[str] = None,
+        limit: Optional[int] = None,
+        sort_by: ExperimentSort = ExperimentSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> List[Experiment]:
+        """
+        List the experiments in a workspace.
+
+        :param workspace: The Beaker workspace name or object. If not specified,
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
+        :param match: Only include experiments matching the text.
+        :param limit: Limit the number of experiments returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
+
+        :raises WorkspaceNotFound: If the workspace doesn't exist.
+        :raises WorkspaceNotSet: If neither ``workspace`` nor
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` are set.
+        :raises BeakerError: Any other :class:`~beaker.exceptions.BeakerError` type that can occur.
+        :raises RequestException: Any other exception that can occur when contacting the
+            Beaker server.
+        """
+        return list(
+            self.iter_experiments(
+                workspace=workspace,
+                match=match,
+                limit=limit,
+                sort_by=sort_by,
+                descending=descending,
+                cursor=cursor,
+            )
+        )
+
+    def iter_datasets(
+        self,
+        workspace: Optional[Union[str, Workspace]] = None,
+        *,
         match: Optional[str] = None,
         results: Optional[bool] = None,
         uncommitted: Optional[bool] = None,
         limit: Optional[int] = None,
-    ) -> List[Dataset]:
+        sort_by: DatasetSort = DatasetSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> Generator[Dataset, None, None]:
         """
-        List the datasets in a workspace.
+        Iterate over the datasets in a workspace.
 
         :param workspace: The Beaker workspace name, or object. If not specified,
             :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
@@ -386,6 +529,9 @@ class WorkspaceClient(ServiceClient):
         :param results: Only include/exclude experiment result datasets.
         :param uncommitted: Only include/exclude uncommitted datasets.
         :param limit: Limit the number of datasets returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
 
         :raises WorkspaceNotFound: If the workspace doesn't exist.
         :raises WorkspaceNotSet: If neither ``workspace`` nor
@@ -395,18 +541,22 @@ class WorkspaceClient(ServiceClient):
             Beaker server.
         """
         workspace_name = self.resolve_workspace(workspace, read_only_ok=True).full_name
-        datasets: List[Dataset] = []
-        cursor: Optional[str] = None
-        query: Dict[str, str] = {}
+        query: Dict[str, str] = {
+            "field": str(sort_by),
+            "order": "descending" if descending else "ascending",
+            "cursor": format_cursor(cursor),
+        }
         if match is not None:
             query["q"] = match
         if results is not None:
             query["results"] = str(results).lower()
         if uncommitted is not None:
             query["committed"] = str(not uncommitted).lower()
+        if limit:
+            query["limit"] = str(limit)
 
+        count = 0
         while True:
-            query["cursor"] = cursor or ""
             page = DatasetsPage.from_json(
                 self.request(
                     f"workspaces/{self.url_quote(workspace_name)}/datasets",
@@ -417,15 +567,61 @@ class WorkspaceClient(ServiceClient):
                     },
                 ).json()
             )
-            datasets.extend(page.data)
-            cursor = page.next_cursor
-            if not cursor:
-                break
-            if limit is not None and len(datasets) >= limit:
-                datasets = datasets[:limit]
+
+            for dataset in page.data:
+                count += 1
+                yield dataset
+                if limit is not None and count >= limit:
+                    return
+
+            query["cursor"] = page.next_cursor or page.next  # type: ignore
+            if not query["cursor"]:
                 break
 
-        return datasets
+    def datasets(
+        self,
+        workspace: Optional[Union[str, Workspace]] = None,
+        *,
+        match: Optional[str] = None,
+        results: Optional[bool] = None,
+        uncommitted: Optional[bool] = None,
+        limit: Optional[int] = None,
+        sort_by: DatasetSort = DatasetSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> List[Dataset]:
+        """
+        List the datasets in a workspace.
+
+        :param workspace: The Beaker workspace name, or object. If not specified,
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
+        :param match: Only include datasets matching the text.
+        :param results: Only include/exclude experiment result datasets.
+        :param uncommitted: Only include/exclude uncommitted datasets.
+        :param limit: Limit the number of datasets returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
+
+        :raises WorkspaceNotFound: If the workspace doesn't exist.
+        :raises WorkspaceNotSet: If neither ``workspace`` nor
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` are set.
+        :raises BeakerError: Any other :class:`~beaker.exceptions.BeakerError` type that can occur.
+        :raises RequestException: Any other exception that can occur when contacting the
+            Beaker server.
+        """
+        return list(
+            self.iter_datasets(
+                workspace=workspace,
+                match=match,
+                results=results,
+                uncommitted=uncommitted,
+                limit=limit,
+                sort_by=sort_by,
+                descending=descending,
+                cursor=cursor,
+            )
+        )
 
     def secrets(self, workspace: Optional[Union[str, Workspace]] = None) -> List[Secret]:
         """
@@ -453,12 +649,26 @@ class WorkspaceClient(ServiceClient):
             ).json()["data"]
         ]
 
-    def groups(self, workspace: Optional[Union[str, Workspace]] = None) -> List[Group]:
+    def iter_groups(
+        self,
+        workspace: Optional[Union[str, Workspace]] = None,
+        *,
+        match: Optional[str] = None,
+        limit: Optional[int] = None,
+        sort_by: GroupSort = GroupSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> Generator[Group, None, None]:
         """
-        List groups in a workspace.
+        Iterate over groups in a workspace.
 
         :param workspace: The Beaker workspace name, or object. If not specified,
             :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
+        :param match: Only include groups matching the text.
+        :param limit: Limit the number of groups returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
 
         :raises WorkspaceNotFound: If the workspace doesn't exist.
         :raises WorkspaceNotSet: If neither ``workspace`` nor
@@ -468,16 +678,77 @@ class WorkspaceClient(ServiceClient):
             Beaker server.
         """
         workspace_name = self.resolve_workspace(workspace, read_only_ok=True).full_name
-        return [
-            Group.from_json(d)
-            for d in self.request(
-                f"workspaces/{self.url_quote(workspace_name)}/groups",
-                method="GET",
-                exceptions_for_status={
-                    404: WorkspaceNotFound(self._not_found_err_msg(workspace_name))
-                },
-            ).json()["data"]
-        ]
+        query: Dict[str, str] = {
+            "field": str(sort_by),
+            "order": "descending" if descending else "ascending",
+            "cursor": format_cursor(cursor),
+        }
+        if match is not None:
+            query["q"] = match
+        if limit:
+            query["limit"] = str(limit)
+
+        count = 0
+        while True:
+            page = GroupsPage.from_json(
+                self.request(
+                    f"workspaces/{self.url_quote(workspace_name)}/groups",
+                    method="GET",
+                    query=query,
+                    exceptions_for_status={
+                        404: WorkspaceNotFound(self._not_found_err_msg(workspace_name))
+                    },
+                ).json()
+            )
+
+            for group in page.data:
+                count += 1
+                yield group
+                if limit is not None and count >= limit:
+                    return
+
+            query["cursor"] = page.next_cursor or page.next  # type: ignore
+            if not query["cursor"]:
+                break
+
+    def groups(
+        self,
+        workspace: Optional[Union[str, Workspace]] = None,
+        *,
+        match: Optional[str] = None,
+        limit: Optional[int] = None,
+        sort_by: GroupSort = GroupSort.created,
+        descending: bool = True,
+        cursor: int = 0,
+    ) -> List[Group]:
+        """
+        List groups in a workspace.
+
+        :param workspace: The Beaker workspace name, or object. If not specified,
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
+        :param match: Only include groups matching the text.
+        :param limit: Limit the number of groups returned.
+        :param sort_by: The field to sort the results by.
+        :param descending: Order the results in descending order according to the ``sort_by`` field.
+        :param cursor: Set the starting cursor for the query. You can use this to paginate the results.
+
+        :raises WorkspaceNotFound: If the workspace doesn't exist.
+        :raises WorkspaceNotSet: If neither ``workspace`` nor
+            :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` are set.
+        :raises BeakerError: Any other :class:`~beaker.exceptions.BeakerError` type that can occur.
+        :raises RequestException: Any other exception that can occur when contacting the
+            Beaker server.
+        """
+        return list(
+            self.iter_groups(
+                workspace=workspace,
+                match=match,
+                limit=limit,
+                sort_by=sort_by,
+                descending=descending,
+                cursor=cursor,
+            )
+        )
 
     def get_permissions(
         self, workspace: Optional[Union[str, Workspace]] = None
@@ -621,11 +892,13 @@ class WorkspaceClient(ServiceClient):
     def clear(
         self,
         workspace: Optional[Union[str, Workspace]] = None,
+        *,
         groups: bool = True,
         experiments: bool = True,
         images: bool = True,
         datasets: bool = True,
         secrets: bool = True,
+        older_than: Optional[datetime] = None,
     ):
         """
         Remove groups, experiments, images, datasets, and secrets from a workspace.
@@ -637,6 +910,7 @@ class WorkspaceClient(ServiceClient):
         :param images: Whether to delete images.
         :param datasets: Whether to delete datasets.
         :param secrets: Whether to delete secrets.
+        :param older_than: Only delete objects created before this date.
 
         :raises WorkspaceNotFound: If the workspace doesn't exist.
         :raises WorkspaceNotSet: If neither ``workspace`` nor
@@ -647,36 +921,50 @@ class WorkspaceClient(ServiceClient):
         """
         import concurrent.futures
 
+        def should_delete(created: Optional[datetime]) -> bool:
+            if older_than is None or created is None:
+                return True
+            if any([dt.tzinfo is None for dt in (created, older_than)]):
+                return created.replace(tzinfo=None) < older_than.replace(tzinfo=None)
+            else:
+                return created < older_than
+
         deletion_counts: Dict[str, int] = defaultdict(int)
         with concurrent.futures.ThreadPoolExecutor() as executor:
             deletion_futures = []
 
             if groups:
-                for group in self.groups(workspace):
+                for group in filter(lambda x: should_delete(x.created), self.groups(workspace)):
                     future = executor.submit(self.beaker.group.delete, group)
                     deletion_futures.append(future)
                     deletion_counts["groups_deleted"] += 1
 
             if experiments:
-                for experiment in self.experiments(workspace):
+                for experiment in filter(
+                    lambda x: should_delete(x.created), self.iter_experiments(workspace)
+                ):
                     future = executor.submit(self.beaker.experiment.delete, experiment)
                     deletion_futures.append(future)
                     deletion_counts["experiments_deleted"] += 1
 
             if images:
-                for image in self.images(workspace):
+                for image in filter(
+                    lambda x: should_delete(x.committed), self.iter_images(workspace)
+                ):
                     future = executor.submit(self.beaker.image.delete, image)
                     deletion_futures.append(future)
                     deletion_counts["images_deleted"] += 1
 
             if datasets:
-                for dataset in self.datasets(workspace):
+                for dataset in filter(
+                    lambda x: should_delete(x.created), self.iter_datasets(workspace)
+                ):
                     future = executor.submit(self.beaker.dataset.delete, dataset)
                     deletion_futures.append(future)
                     deletion_counts["datasets_deleted"] += 1
 
             if secrets:
-                for secret in self.secrets(workspace):
+                for secret in filter(lambda x: should_delete(x.created), self.secrets(workspace)):
                     future = executor.submit(self.beaker.secret.delete, secret, workspace)
                     deletion_futures.append(future)
                     deletion_counts["secrets_deleted"] += 1